disk_backend: remove GetLbaStatus
There is no support in any backend for querying LBA status, and the
trait for querying it is not really suitable for use with async disks.
All devices that opt into supporting LBA status queries just end up
using the default implementation, which reports that all blocks are
fully mapped.

Perhaps a future change will actually support these queries properly.
For now, move this functionality entirely into the SCSI emulator.
